Prepare files for decomp
