+++ config/symbols.us.stnz0.txt
@@ -101,4 +101,27 @@ g_LayoutObjVertical = 0x801CAA78;
 g_LayoutObjPosHorizontal = 0x801CAA7C;
 g_LayoutObjPosVertical = 0x801CAA80;
 g_Dialogue = 0x801CB688;
+g_Dialogue_startX = 0x801CB68C;
+g_Dialogue_nextLineY = 0x801CB68E;
+g_Dialogue_startY = 0x801CB690;
+g_Dialogue_nextCharX = 0x801CB692;
+g_Dialogue_nextLineX = 0x801CB694;
+g_Dialogue_nextCharY = 0x801CB696;
+g_Dialogue_portraitAnimTimer = 0x801CB698;
+g_Dialogue_unk12 = 0x801CB69A;
+g_Dialogue_clutIndex = 0x801CB69C;
+g_Dialogue_nextCharTimer = 0x801CB69E;
+g_Dialogue_unk17 = 0x801CB69F;
+g_Dialogue_prim_0 = 0x801CB6A0;
+g_Dialogue_prim_1 = 0x801CB6A4;
+g_Dialogue_prim_2 = 0x801CB6A8;
+g_Dialogue_prim_3 = 0x801CB6AC;
+g_Dialogue_prim_4 = 0x801CB6B0;
+g_Dialogue_prim_5 = 0x801CB6B4;
+g_Dialogue_primIndex_0 = 0x801CB6B8;
+g_Dialogue_primIndex_1 = 0x801CB6BC;
+g_Dialogue_primIndex_2 = 0x801CB6C0;
+g_Dialogue_unk3C = 0x801CB6C4;
+g_Dialogue_timer = 0x801CB6C6;
+g_Dialogue_unk40 = 0x801CB6C8;
 g_ItemIconSlots = 0x801CB740;
